mild refactor, significant improvement
moved repo allocation out of Commit constructor. now I'm not allocating the repo
5,000+ times when I analyze Aaron Patterson's commit history. the result is a
speedup from about 92 or 93 seconds to 20 or 21. compared to the original 3.5
minutes time, which is about 210 seconds, that is an improvement of an order of
magnitude.

the Clojure version is still a little more than 2x faster, but this is a big
improvement for the Ruby version nonetheless.